backend/requirements.txt
# ─── Petpooja AI Copilot — Backend Dependencies ───
# All local, no external API SDKs needed

# Web framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# Database (Supabase PostgreSQL)
sqlalchemy==2.0.36
psycopg2-binary==2.9.10

# Data processing
pandas==2.2.3
numpy==2.2.1

# Association rule mining (combo engine)
mlxtend==0.23.2

# Voice — local STT
faster-whisper==1.1.0

# Fuzzy text matching
rapidfuzz==3.11.0

# Hindi/Hinglish text processing
indic-transliteration==2.4.1

# Audio handling
soundfile==0.13.1
pydub==0.25.1

# Utilities
python-dotenv==1.0.1
